+++ src/main/kotlin/com/nbadal/ktlint/KtlintConfig.kt
@@ -9,7 +9,7 @@ import javax.swing.JComponent
 class KtlintConfig(private val project: Project) : SearchableConfigurable {
 
     private val configStorage = KtlintConfigStorage.instance(project)
-    private val form = KtlintConfigForm(configStorage)
+    private val form = KtlintConfigForm(project, configStorage)
 
     override fun createComponent(): JComponent = form.createComponent()