+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Changed
 - Update ktlint to 0.39.0
+- Opt-out of kotlint stdlib inclusion
 
 ### Deprecated