+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## [Unreleased]
 ### Added
 - Plugin setting for disabled rules
+- Add support for .editorconfig
 
 ### Changed
 - Update ktlint to 0.39.0